refactor: 카카오 지도를 shared로 추출
