+++ apps/web/src/shared/lib/index.ts
@@ -4,3 +4,4 @@ export * from './daum-postcode';
 export * from './image-save';
 export * from './format';
 export { mswHttp } from './msw';
+export * from './kakao-map';
