+++ apps/web/src/shared/lib/kakao-map/index.ts
@@ -1,0 +1,1 @@
+export { loadKakaoMapSDK } from './kakao-map';
