+++ apps/web/src/shared/lib/kakao-map/kakao-map.ts
@@ -1,0 +1,49 @@
+import { KAKAO_JS_KEY } from '@/shared/storage';
+
+declare global {
+  interface Window {
+    kakao: {
+      maps: {
+        load: (callback: () => void) => void;
+        Map: new (
+          container: HTMLElement,
+          options: { center: unknown; level: number },
+        ) => unknown;
+        LatLng: new (lat: number, lng: number) => unknown;
+        Marker: new (options: { map: unknown; position: unknown }) => unknown;
+        services: {
+          Geocoder: new () => {
+            addressSearch: (
+              address: string,
+              callback: (
+                result: { x: string; y: string }[],
+                status: string,
+              ) => void,
+            ) => void;
+          };
+          Status: { OK: string };
+        };
+      };
+    };
+  }
+}
+
+const KAKAO_MAP_SDK_URL = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${KAKAO_JS_KEY}&libraries=services&autoload=false`;
+
+export const loadKakaoMapSDK = (): Promise<void> => {
+  return new Promise((resolve, reject) => {
+    if (window.kakao?.maps) {
+      window.kakao.maps.load(() => resolve());
+      return;
+    }
+
+    const script = document.createElement('script');
+    script.src = KAKAO_MAP_SDK_URL;
+    script.async = true;
+    script.onload = () => {
+      window.kakao.maps.load(() => resolve());
+    };
+    script.onerror = () => reject(new Error('카카오 지도 SDK 로드 실패'));
+    document.head.appendChild(script);
+  });
+};
